requirements.txt
aif360==0.4.0
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.9
cloudpickle==2.0.0
configparser==5.2.0
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.4.2
fonttools==4.28.5
gitdb==4.0.9
GitPython==3.1.24
googledrivedownloader==0.4
h5py
huggingface-hub==0.2.1
idna==3.3
iniconfig==1.1.1
isodate==0.6.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
littleutils==0.2.2
MarkupSafe==2.0.1
matplotlib==3.5.1
memory-profiler==0.60.0
networkx==2.6.3
numpy==1.20.2
ogb==1.3.2
outdated==0.2.1
packaging==21.3
pandas==1.3.0
pathtools==0.1.2
Pillow==8.4.0
pluggy==1.0.0
promise==2.3
protobuf==3.19.1
psutil==5.9.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
rdflib==6.1.1
regex==2021.11.10
requests==2.26.0
sacremoses==0.0.46
scikit-learn==1.0.2
scipy==1.5.4
sentry-sdk==1.5.1
shap==0.40.0
shortuuid==1.0.8
six==1.16.0
slicer==0.0.7
smmap==5.0.0
subprocess32==3.5.4
tempeh==0.1.12
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.0
torch-geometric==2.0.3
torch-scatter==2.0.9
torch-sparse==0.6.12
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.62.3
transformers==4.15.0
urllib3==1.26.7
wandb==0.12.9
wilds==2.0.0
yacs==0.1.8
yaspin==2.1.0
